+++ woodwork/column_accessor.py
@@ -8,8 +8,7 @@ from woodwork.accessor_utils import (
     _check_column_schema,
     _is_dataframe,
     _is_series,
-    init_series,
-    is_schema_valid
+    init_series
 )
 from woodwork.column_schema import ColumnSchema
 from woodwork.exceptions import (
@@ -258,18 +257,18 @@ class WoodworkColumnAccessor:
                                                                                           'Series')
                         warnings.warn(warning_message, TypingInfoMismatchWarning)
                 elif _is_dataframe(result):
-                    # Initialize Woodwork if a valid table schema can be created from the original column schema
+                    # Initialize Woodwork with a partial schema
                     col_schema = self.schema
-                    table_schema = TableSchema(column_names=[self.name],
-                                               logical_types={self.name: col_schema.logical_type},
-                                               semantic_tags={self.name: col_schema.semantic_tags},
-                                               column_metadata={self.name: col_schema.metadata},
-                                               use_standard_tags={self.name: col_schema.use_standard_tags},
-                                               column_descriptions={self.name: col_schema.description},
-                                               column_origins={self.name: col_schema.origin},
+                    col_name = self.name or result.columns.to_list()[0]
+                    table_schema = TableSchema(column_names=[col_name],
+                                               logical_types={col_name: col_schema.logical_type},
+                                               semantic_tags={col_name: col_schema.semantic_tags},
+                                               column_metadata={col_name: col_schema.metadata},
+                                               use_standard_tags={col_name: col_schema.use_standard_tags},
+                                               column_descriptions={col_name: col_schema.description},
+                                               column_origins={col_name: col_schema.origin},
                                                validate=False)
-                    if is_schema_valid(result, table_schema):
-                        result.ww.init(schema=table_schema)
+                    result.ww.init_with_partial_schema(table_schema)
                 # Always return the results of the Series operation whether or not Woodwork is initialized
                 return result
             return wrapper